requirements.txt

fastapi==0.115.5
uvicorn==0.32.0
python-multipart==0.0.17
pydantic==2.9.2
pydantic_core==2.23.4
starlette==0.41.2
orjson==3.10.12


httpcore==1.0.7
httpx==0.27.2
requests==2.32.3

pandas==2.0.3
numpy==1.24.4
openpyxl==3.1.5
SQLAlchemy==2.0.36
python-dotenv==1.0.1
anyio==4.5.2
sniffio==1.3.1
aiohttp==3.10.11
async-timeout==4.0.3
yarl==1.15.2
frozenlist==1.5.0
multidict==6.1.0


langchain==0.2.17
langchain-core==0.2.43
langchain-openai==0.1.25
langsmith==0.1.146
openai==1.55.1
tiktoken==0.7.0
transformers==4.46.3
torch==2.4.1
huggingface-hub==0.36.0
tokenizers==0.20.3
safetensors==0.5.3
sentence-transformers==3.2.1
sentencepiece==0.2.0
scikit-learn==1.3.2
scipy==1.10.1
joblib==1.4.2